@@ -9,10 +9,11 @@
     "deploy": "docusaurus deploy"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.0.0-alpha.55",
-    "@docusaurus/preset-classic": "^2.0.0-alpha.55",
+    "@docusaurus/core": "^2.0.0-alpha.66",
+    "@docusaurus/preset-classic": "^2.0.0-alpha.61",
+    "@docusaurus/theme-search-algolia": "^2.0.0-alpha.32",
     "classnames": "^2.2.6",
-    "react": "^16.8.4",
+    "react": "^16.13.1",
     "react-dom": "^16.8.4",
     "react-feather": "^2.0.8"
   },
