@@ -57,17 +57,17 @@
     "graphql-tag": "^2.6.1",
     "i18n-react": "^0.6.4",
     "json-loader": "^0.5.4",
+    "node-sass": "^4.14.1",
     "react": "16.4.2",
     "react-apollo": "^2.1.9",
     "react-dom": "16.4.2",
     "react-transition-group": "^2.4.0",
+    "sass-loader": "^6.0.6",
     "source-map-loader": "^0.2.3",
     "style-loader": "^0.16.1",
     "subscriptions-transport-ws": "^0.9.14",
     "uglifyjs-webpack-plugin": "^2.2.0",
     "url-loader": "^1.0.1",
-    "node-sass": "^4.14.1",
-    "sass-loader": "^6.0.6",
     "webpack": "4.16.3",
     "webpack-cli": "^3.1.0",
     "webpack-merge": "^4.1.3",
@@ -77,7 +77,7 @@
     "git-repo-info": "^2.1.0",
     "nodemon": "^1.11.0",
     "ts-node": "^8.3.0",
-    "typescript": "^2.9.2"
+    "typescript": "^3.5.2"
   },
   "engines": {
     "node": ">=10.x.x"
