@@ -49,15 +49,25 @@ describe('Test configs mutations', () => {
   });
 
   test('Activate installation', async () => {
-    const mock = sinon.stub(utils, 'sendRequest').callsFake(() => {
-      return Promise.resolve('ok');
-    });
-
     const qry = `
       mutation configsActivateInstallation($token: String! $hostname: String!) {
         configsActivateInstallation(token: $token hostname: $hostname)
       }
     `;
+
+    try {
+      await graphqlRequest(qry, 'configsActivateInstallation', {
+        hostname: 'localhost',
+        token: 'token'
+      });
+    } catch (e) {
+      console.log('e: ', e);
+      expect(e[0].message).toBe('');
+    }
+
+    const mock = sinon.stub(utils, 'sendRequest').callsFake(() => {
+      return Promise.resolve('ok');
+    });
 
     const response = await graphqlRequest(qry, 'configsActivateInstallation', {
       hostname: 'localhost',
