@@ -128,18 +128,6 @@ describe('engage messages model tests', () => {
     expect(messagesCounts).toBe(0);
   });
 
-  test('remove a message: can not remove manual message ', async () => {
-    expect.assertions(1);
-
-    const manualMessage = await engageMessageFactory({ kind: 'manual' });
-
-    try {
-      await EngageMessages.removeEngageMessage(manualMessage._id);
-    } catch (e) {
-      expect(e.message).toBe('Can not remove manual message');
-    }
-  });
-
   test('Engage message set live', async () => {
     await EngageMessages.engageMessageSetLive(_message._id);
     const message = await EngageMessages.findOne({ _id: _message._id });
@@ -351,21 +339,71 @@ describe('createConversation', () => {
 
     expect(message1.isCustomerRead).toBe(true);
     expect(message2.isCustomerRead).toBe(true);
+
+    // message is equal with previous one =====================
+    const engageMessage = await engageMessageFactory({
+      messenger: {
+        brandId: 'brandId',
+        kind: 'chat',
+        sentAs: 'snippet',
+        content: 'content'
+      },
+      customerIds: [_customer._id],
+      fromUserId: user._id
+    });
+
+    await conversationMessageFactory({
+      conversationId: conversation._id,
+      engageData: engageDataFactory({
+        messageId: engageMessage._id,
+        brandId: 'brandId',
+        kind: 'chat',
+        sentAs: 'snippet',
+        content: 'content',
+        fromUserId: user._id
+      }),
+      userId: user._id,
+      customerId: _customer._id
+    });
+
+    response = await EngageMessages.createOrUpdateConversationAndMessages({
+      customer: _customer,
+      integration: _integration,
+      user,
+      replacedContent,
+      engageData: {
+        rules: [],
+        messageId: engageMessage._id,
+        brandId: 'brandId',
+        content: 'content',
+        fromUserId: user._id,
+        kind: 'chat',
+        sentAs: 'snippet'
+      }
+    });
+
+    expect(response).toBe(null);
   });
 });
 
-describe('createVisitorMessages', () => {
+describe('createVisitorOrCustomerMessages', () => {
   let _user: IUserDocument;
   let _brand: IBrandDocument;
   let _customer: ICustomerDocument;
   let _integration: IIntegrationDocument;
+  let _visitor: ICustomerDocument;
   let mock;
 
   beforeEach(async () => {
     // Creating test data
     _customer = await customerFactory({
       firstName: 'firstName',
-      lastName: 'lastName'
+      lastName: 'lastName',
+      state: 'customer'
+    });
+
+    _visitor = await customerFactory({
+      state: 'visitor'
     });
 
     mock = sinon.stub(events, 'getNumberOfVisits').callsFake(() => {
@@ -402,9 +440,31 @@ describe('createVisitorMessages', () => {
 
     // invalid from user id
     await engageMessageFactory({
-      kind: 'visitorAuto',
+      kind: 'auto',
       userId: 'invalid',
       isLive: true,
+      messenger: {
+        brandId: _brand._id,
+        content: 'hi,{{ customer.firstName }} {{ customer.lastName }}'
+      }
+    });
+
+    await engageMessageFactory({
+      kind: 'visitorAuto',
+      userId: _user._id,
+      isLive: true,
+      customerIds: [_visitor.id],
+      messenger: {
+        brandId: _brand._id,
+        content: 'hi,{{ customer.firstName }} {{ customer.lastName }}'
+      }
+    });
+
+    await engageMessageFactory({
+      kind: 'manual',
+      userId: _user._id,
+      isLive: true,
+      customerIds: [_customer.id],
       messenger: {
         brandId: _brand._id,
         content: 'hi,{{ customer.firstName }} {{ customer.lastName }}'
@@ -444,13 +504,45 @@ describe('createVisitorMessages', () => {
       })
     });
 
+    const fromUser = await userFactory({});
+
+    const eng = await engageMessageFactory({
+      messenger: {
+        brandId: _brand._id,
+        kind: 'chat',
+        sentAs: 'snippet',
+        content: 'content',
+        rules: [
+          {
+            kind: 'currentPageUrl',
+            text: 'Current page url',
+            condition: 'contains',
+            value: 'index'
+          }
+        ]
+      },
+      customerIds: [_customer._id],
+      isLive: true,
+      method: 'messenger',
+      kind: 'auto',
+      userId: fromUser._id
+    });
+
+    await conversationMessageFactory({
+      customerId: _customer._id,
+      isCustomerRead: false,
+      engageData: engageDataFactory({
+        messageId: eng._id,
+        fromUserId: eng.fromUserId
+      })
+    });
     // main call
-    const msgs = await EngageMessages.createVisitorMessages({
+    const msgs = await EngageMessages.createVisitorOrCustomerMessages({
       brand: _brand,
       customer: _customer,
       integration: _integration,
       browserInfo: {
-        url: '/page'
+        url: '/index'
       }
     });
 
@@ -480,15 +572,108 @@ describe('createVisitorMessages', () => {
 
     expect(message._id).toBeDefined();
     expect(message.content).toBe(content);
+  });
 
-    // count of unread conversation messages created by engage must be zero
-    const convEngageMessages = await Messages.find({
-      customerId: _customer._id,
-      isCustomerRead: false,
-      engageData: { $exists: true }
+  test('manual message test', async () => {
+    const customer1 = await customerFactory({
+      state: 'customer',
+      firstName: 'firstName',
+      lastName: 'lastName'
+    });
+    const customer2 = await customerFactory({ state: 'customer' });
+
+    const engageMessage = await engageMessageFactory({
+      kind: 'manual',
+      userId: _user._id,
+      isLive: true,
+      customerIds: [customer1.id],
+      messenger: {
+        brandId: _brand._id,
+        content: 'hi,{{ customer.firstName }} {{ customer.lastName }}'
+      }
     });
 
-    expect(convEngageMessages.length).toBe(0);
+    await EngageMessages.createVisitorOrCustomerMessages({
+      brand: _brand,
+      customer: customer2,
+      integration: _integration,
+      browserInfo: {
+        url: '/index'
+      }
+    });
+
+    await EngageMessages.createVisitorOrCustomerMessages({
+      brand: _brand,
+      customer: customer1,
+      integration: _integration,
+      browserInfo: {
+        url: '/index'
+      }
+    });
+
+    const conversation1 = await Conversations.findOne({
+      customerId: customer1._id
+    });
+
+    const conversation2 = await Conversations.findOne({
+      customerId: customer2._id
+    });
+
+    if (!conversation1) {
+      throw new Error('conversation not found');
+    }
+
+    const content = `hi,${_customer.firstName || ''} ${_customer.lastName ||
+      ''}`;
+
+    expect(conversation1?._id).toBeDefined();
+    expect(conversation2).toBeNull();
+    expect(conversation1?.content).toBe(content);
+    expect(conversation1?.customerId).toBe(customer1._id);
+    // expect(conversation.integrationId).toBe(_integration._id);
+
+    const message = await Messages.findOne({
+      conversationId: conversation1?._id
+    });
+
+    if (!message) {
+      throw new Error('message not found');
+    }
+
+    expect(message._id).toBeDefined();
+    expect(message.content).toBe(content);
+    expect(message.engageData).toBeDefined();
+    expect(message.engageData?.messageId).toBe(engageMessage._id);
+  });
+
+  test('manual or auto message: customer does not exist test', async () => {
+    const customer = await customerFactory({ state: 'customer' });
+
+    await engageMessageFactory({
+      kind: 'manual',
+      userId: _user._id,
+      isLive: true,
+      tagIds: [(await tagsFactory({}))._id],
+      messenger: {
+        brandId: _brand._id,
+        content: 'hi,{{ customer.firstName }} {{ customer.lastName }}'
+      }
+    });
+
+    await EngageMessages.createVisitorOrCustomerMessages({
+      brand: _brand,
+      customer,
+      integration: _integration,
+      browserInfo: {
+        url: '/index'
+      }
+    });
+
+    const conversation = await Conversations.findOne({
+      customerId: customer._id
+    });
+
+    expect(conversation).toBeNull();
   });
 
   const browserLanguageRule = {
