@@ -12,8 +12,42 @@ export const types = `
     _id: String!
     name: String
     isPrivate: Boolean
+    boardId: String
+    memberIds: [String]
 
     calendars: [Calendar]
+  }
+
+  type CalendarBoard {
+    _id: String!
+    name: String
+
+    groups: [CalendarGroup]
+  }
+
+  type NylasCalendar {
+    _id: String
+    providerCalendarId: String
+    accountUid: String
+    name: String
+    description: String
+    readOnly: Boolean
+  }
+
+  type FullCalendar {
+    _id: String!
+    name: String
+    color: String
+    accountId: String
+
+    calendars: [NylasCalendar]
+  }
+
+ input Participant {
+    name: String
+    email: String
+    status: String
+    comment: String
   }
 `;
 
@@ -23,28 +57,39 @@ const eventParams = `
   title: String!,
   description: String,
   start: String,
-  end: String
+  end: String,
+  
+  participants: [Participant]
+  memberIds: [String]
 `;
 
 const commonParams = `
   groupId: String!,
-  name: String!,
   color: String,
 `;
 
 const commonGroupParams = `
   name: String!,
+  boardId: String!,
   isPrivate: Boolean,
-  assignedUserIds: [String]
+  memberIds: [String],
 `;
 
 export const queries = `
-  calendarGroups: [CalendarGroup]
+  calendarBoards: [CalendarBoard]
+  calendarBoardCounts: Int
+  calendarBoardGetLast: CalendarBoard
+  calendarBoardDetail(_id: String!): CalendarBoard
+
+  calendarGroups(boardId: String): [CalendarGroup]
   calendarGroupCounts: Int
   calendarGroupGetLast: CalendarGroup
   calendarGroupDetail(_id: String!): CalendarGroup
+
   calendars(groupId: String, page: Int, perPage: Int): [Calendar]
   calendarDetail(_id: String!): Calendar
+
+  calendarAccounts(groupId: String): [FullCalendar]
 `;
 
 export const mutations = `
@@ -59,4 +104,8 @@ export const mutations = `
   calendarGroupsAdd(${commonGroupParams}): CalendarGroup
   calendarGroupsEdit(_id: String!, ${commonGroupParams}): CalendarGroup
   calendarGroupsDelete(_id: String!): JSON
+
+  calendarBoardsAdd(name: String!): CalendarBoard
+  calendarBoardsEdit(_id: String!, name: String): CalendarBoard
+  calendarBoardsDelete(_id: String!): JSON
 `;
