@@ -29,7 +29,12 @@ const connectProviderToNylas = async (uid: string, integrationId?: string) => {
     throw new Error(`Refresh token not found ${uid}`);
   }
 
-  const [email, refreshToken, kind] = providerCredential.split(',');
+  const [
+    email,
+    refreshToken,
+    kind,
+    googleAccessToken
+  ] = providerCredential.split(',');
 
   if (integrationId) {
     const isEmailDuplicated = await checkEmailDuplication(email, kind);
@@ -73,12 +78,14 @@ const connectProviderToNylas = async (uid: string, integrationId?: string) => {
         integrationId,
         nylasToken: access_token,
         nylasAccountId: account_id,
-        status: billing_state
+        status: billing_state,
+        googleAccessToken
       });
     } else {
       const newAccount = await Accounts.create({
         kind,
         email,
+        googleAccessToken,
         nylasToken: access_token,
         nylasAccountId: account_id,
         nylasBillingState: 'paid'
@@ -321,7 +328,8 @@ const createIntegration = async ({
   nylasAccountId,
   nylasToken,
   status,
-  kind
+  kind,
+  googleAccessToken
 }: {
   email?: string;
   integrationId: string;
@@ -329,6 +337,7 @@ const createIntegration = async ({
   nylasToken: string;
   status: string;
   kind: string;
+  googleAccessToken?: string;
 }) => {
   if (status === 'cancelled') {
     await enableOrDisableAccount(nylasAccountId, true);
@@ -338,6 +347,7 @@ const createIntegration = async ({
     ...(email ? { email } : {}),
     kind,
     erxesApiId: integrationId,
+    googleAccessToken,
     nylasToken,
     nylasAccountId,
     nylasBillingState: 'paid'
