@@ -1,6 +1,12 @@
 import { Document, Model, model, Schema } from 'mongoose';
 import { field } from '../models/utils';
 
+export interface IRecording {
+  id: string;
+  url?: string;
+  expires?: number;
+}
+
 export interface ICallRecord {
   erxesApiMessageId: string;
   erxesApiConversationId: string;
@@ -8,18 +14,30 @@ export interface ICallRecord {
   kind: string;
   privacy: string;
   status?: string;
-  recordId?: string;
   token?: string;
+  recordings?: IRecording[];
 }
 
 interface ICallRecordDocument extends ICallRecord, Document {
   _id: string;
 }
 
+const recordingSchema = new Schema(
+  {
+    id: String,
+    url: String,
+    expires: Number
+  },
+  {
+    _id: false
+  }
+);
+
 const callRecordSchema = new Schema({
   _id: field({ pkey: true }),
   erxesApiMessageId: String,
   erxesApiConversationId: String,
+  recordings: [recordingSchema],
   roomName: String,
   kind: String,
   privacy: String,
@@ -28,7 +46,6 @@ const callRecordSchema = new Schema({
     type: String,
     default: 'ongoing'
   },
-  recordId: String,
   createdAt: Date
 });
 
