@@ -4,19 +4,59 @@ import Icon from 'modules/common/components/Icon';
 import ModalTrigger from 'modules/common/components/ModalTrigger';
 import React from 'react';
 import { EventContainer, EventContent, EventTitle } from '../styles';
-import { IEvent } from '../types';
+import { IAccount, IEvent } from '../types';
 import { milliseconds } from '../utils';
+import { CalendarConsumer } from './Wrapper';
 
 type Props = {
   event: IEvent;
   showHour: boolean;
-  editEvent: (event: IEvent) => void;
-  deleteEvent: (event: IEvent) => void;
+  editEvent: (event: IEvent, account?: IAccount) => void;
+  deleteEvent: (_id: string, accountId: string) => void;
 };
 
-class Detail extends React.Component<Props> {
+type FinalProps = {
+  color: object;
+  accounts: IAccount[];
+} & Props;
+
+class Detail extends React.Component<FinalProps, { toggle: boolean }> {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      toggle: false
+    };
+  }
+
+  getColor(color: object, accountId: string) {
+    const colorCode = color[accountId];
+
+    return colorCode;
+  }
+
+  getAccount() {
+    const { event, accounts } = this.props;
+
+    for (const account of accounts) {
+      const acc = account.calendars.find(
+        c => c.providerCalendarId === event.providerCalendarId
+      );
+
+      if (acc) {
+        return account;
+      }
+    }
+
+    return;
+  }
+
+  onToggle = () => {
+    this.setState({ toggle: !this.state.toggle });
+  };
+
   render() {
-    const { event, showHour, editEvent, deleteEvent } = this.props;
+    const { event, showHour, editEvent, deleteEvent, color } = this.props;
     const startTime = milliseconds(event.when.start_time);
     const endTime = milliseconds(event.when.end_time);
 
@@ -28,13 +68,20 @@ class Detail extends React.Component<Props> {
 
       const edit = () => {
         closeModal();
-        editEvent(event);
+        editEvent(event, this.getAccount());
       };
 
       const remove = () => {
+        const account = this.getAccount();
         closeModal();
-        deleteEvent(event);
+
+        if (account) {
+          deleteEvent(event.providerEventId, account.accountId);
+        }
       };
+
+      const { toggle } = this.state;
+      const guestCount = event.participants.length;
 
       return (
         <EventContent>
@@ -48,7 +95,23 @@ class Detail extends React.Component<Props> {
           <div>{event.owner}</div>
           <br />
           <Icon icon="users" />
-          <div>{event.participants.length} guests</div>
+          <div>
+            {guestCount} guests &nbsp;
+            {guestCount !== 0 && (
+              <Icon
+                icon={`arrow-${toggle ? 'up' : 'down'}`}
+                onClick={this.onToggle}
+                style={{ cursor: 'pointer' }}
+              />
+            )}
+          </div>
+          {toggle && (
+            <ul>
+              {event.participants.map(p => (
+                <li key={p.email}>{p.name}</li>
+              ))}
+            </ul>
+          )}
           {event.description && (
             <>
               <br /> <Icon icon="book" />
@@ -82,7 +145,10 @@ class Detail extends React.Component<Props> {
         <ModalTrigger
           title={event.title || ''}
           trigger={
-            <EventTitle {...props}>
+            <EventTitle
+              {...props}
+              color={this.getColor(color, event.providerCalendarId)}
+            >
               <Icon icon="check-circle" />
               {dayjs(startTime).format('ha')} &nbsp;
               <b>{event.title}</b>
@@ -95,4 +161,14 @@ class Detail extends React.Component<Props> {
   }
 }
 
-export default Detail;
+const WithConsumer = (props: Props) => {
+  return (
+    <CalendarConsumer>
+      {({ color, accounts }) => (
+        <Detail {...props} color={color} accounts={accounts} />
+      )}
+    </CalendarConsumer>
+  );
+};
+
+export default WithConsumer;
