@@ -6,32 +6,31 @@ import FormGroup from 'modules/common/components/form/Group';
 import Label from 'modules/common/components/form/Label';
 import ControlLabel from 'modules/common/components/form/Label';
 import Icon from 'modules/common/components/Icon';
-import { router } from 'modules/common/utils';
 import Sidebar from 'modules/layout/components/Sidebar';
 import React from 'react';
 import Select from 'react-select-plus';
 import { TYPES } from '../constants';
 import EventForm from '../containers/EventForm';
 import { CalendarController, SidebarWrapper } from '../styles';
-import { ICalendar } from '../types';
+import { IAccount } from '../types';
 import { extractDate } from '../utils';
 
 type Props = {
   dateOnChange: (date: string | Date | undefined) => void;
   currentDate: Date;
   typeOnChange: ({ value, label }: { value: string; label: string }) => void;
+  onChangeCalendarIds: (ids: string[]) => void;
   type: string;
-  calendars: ICalendar[];
   history: any;
   queryParams: any;
   startTime: Date;
   endTime: Date;
-  accountId: string;
+  accounts: IAccount[];
 };
 
 type State = {
-  calendarIds: string[];
   isPopupVisible: boolean;
+  calendarIds: string[];
 };
 
 class LeftSidebar extends React.Component<Props, State> {
@@ -39,9 +38,40 @@ class LeftSidebar extends React.Component<Props, State> {
     super(props);
 
     this.state = {
-      calendarIds: [],
-      isPopupVisible: false
+      isPopupVisible: false,
+      calendarIds: this.getCalendarIds(props.accounts)
     };
+  }
+
+  componentWillReceiveProps(nextProps) {
+    if (nextProps.accounts !== this.props.accounts) {
+      const calendarIds = this.getCalendarIds(nextProps.accounts);
+
+      this.setState({ calendarIds });
+      this.props.onChangeCalendarIds(calendarIds);
+    }
+  }
+
+  componentDidMount() {
+    const { calendarIds } = this.state;
+
+    if (calendarIds.length !== 0) {
+      this.props.onChangeCalendarIds(calendarIds);
+    }
+  }
+
+  getCalendarIds(accounts: IAccount[]) {
+    const calendarIds: string[] = [];
+
+    accounts.map(acc => {
+      calendarIds.push(acc._id);
+
+      return acc.calendars
+        .filter(c => !c.readOnly)
+        .map(cal => calendarIds.push(cal.providerCalendarId));
+    });
+
+    return calendarIds;
   }
 
   onHideModal = () => {
@@ -49,33 +79,6 @@ class LeftSidebar extends React.Component<Props, State> {
       isPopupVisible: !this.state.isPopupVisible
     });
   };
-
-  setCalendarIds = () => {
-    const { queryParams, calendars, history } = this.props;
-
-    if (!queryParams.calendarIds) {
-      const calendarIds = calendars.map(c => c.providerCalendarId);
-
-      this.setState({ calendarIds });
-      router.setParams(history, { calendarIds: calendarIds.toString() }, true);
-    }
-  };
-
-  componentDidUpdate() {
-    this.setCalendarIds();
-  }
-
-  componentDidMount() {
-    const queryParams = this.props.queryParams;
-
-    this.setCalendarIds();
-
-    if (queryParams.calendarIds) {
-      this.setState({
-        calendarIds: queryParams.calendarIds.split(',')
-      });
-    }
-  }
 
   renderOptions = (list: string[]) => {
     return list.map(item => ({ value: item, label: item.toUpperCase() }));
@@ -111,25 +114,96 @@ class LeftSidebar extends React.Component<Props, State> {
       calendarIds.push(calendarId);
     } else {
       const index = calendarIds.indexOf(calendarId);
-      calendarIds.splice(index, 1);
+
+      if (index >= 0) {
+        calendarIds.splice(index, 1);
+      }
     }
 
     this.setState({ calendarIds });
 
-    router.setParams(this.props.history, {
-      calendarIds: calendarIds.toString()
+    this.props.onChangeCalendarIds(calendarIds);
+  };
+
+  toggleAccountCheckbox = (
+    account: IAccount,
+    e: React.FormEvent<HTMLElement>
+  ) => {
+    const checked = (e.target as HTMLInputElement).checked;
+    let calendarIds = this.state.calendarIds || [];
+    const providerCalendarIds = account.calendars.map(
+      c => c.providerCalendarId
+    );
+
+    const ids = [account._id, ...providerCalendarIds];
+
+    if (checked) {
+      calendarIds = calendarIds.concat(ids);
+    } else {
+      ids.map(id => {
+        const index = calendarIds.indexOf(id);
+        return index >= 0 && calendarIds.splice(index, 1);
+      });
+    }
+
+    this.setState({ calendarIds });
+
+    this.props.onChangeCalendarIds(calendarIds);
+  };
+
+  renderCalendars = calendars => {
+    const { calendarIds } = this.state;
+
+    return calendars.map(calendar => {
+      const calendarId = calendar.providerCalendarId;
+
+      return (
+        <div key={calendar._id}>
+          &nbsp; &nbsp; &nbsp;
+          <FormControl
+            key={calendar._id}
+            className="toggle-message"
+            componentClass="checkbox"
+            onChange={this.toggleCheckbox.bind(this, calendarId)}
+            checked={calendarIds.includes(calendarId)}
+          >
+            {calendar.name}
+          </FormControl>
+        </div>
+      );
     });
   };
 
+  renderAccounts = () => {
+    return (
+      <FormGroup>
+        <ControlLabel>Calendars</ControlLabel>
+        <br />
+        <br />
+
+        {this.props.accounts.map(account => {
+          return (
+            <div key={account._id}>
+              <FormControl
+                className="toggle-message"
+                componentClass="checkbox"
+                onChange={this.toggleAccountCheckbox.bind(this, account)}
+                checked={this.state.calendarIds.includes(account._id)}
+              >
+                <Icon icon={'circle'} style={{ color: account.color }} /> &nbsp;
+                {account.name}
+              </FormControl>
+              {this.renderCalendars(account.calendars)}
+              <br />
+            </div>
+          );
+        })}
+      </FormGroup>
+    );
+  };
+
   render() {
-    const {
-      type,
-      typeOnChange,
-      currentDate,
-      dateOnChange,
-      calendars
-    } = this.props;
-    const { calendarIds } = this.state;
+    const { type, typeOnChange, currentDate, dateOnChange } = this.props;
 
     return (
       <Sidebar>
@@ -177,29 +251,7 @@ class LeftSidebar extends React.Component<Props, State> {
             />
           </FormGroup>
 
-          <FormGroup>
-            <ControlLabel>Calendars</ControlLabel>
-            <br />
-            <br />
-
-            {calendars.map(calendar => {
-              const calendarId = calendar.providerCalendarId;
-
-              return (
-                <div key={calendar._id}>
-                  <FormControl
-                    key={calendar._id}
-                    className="toggle-message"
-                    componentClass="checkbox"
-                    onChange={this.toggleCheckbox.bind(this, calendarId)}
-                    checked={calendarIds.includes(calendarId)}
-                  >
-                    {calendar.name}
-                  </FormControl>
-                </div>
-              );
-            })}
-          </FormGroup>
+          {this.renderAccounts()}
 
           <FormGroup>
             <Button
