@@ -8,15 +8,14 @@ import React from 'react';
 import { graphql } from 'react-apollo';
 import Event from '../components/Event';
 import { mutations, subscriptions } from '../graphql';
-import { IEvent } from '../types';
 
 type Props = {
   type: string;
   currentDate: Date;
-  accountId: string;
   queryParams: any;
   startTime: Date;
   endTime: Date;
+  calendarIds: string[];
 };
 
 type FinalProps = {
@@ -38,6 +37,10 @@ class EventContainer extends React.Component<FinalProps, {}> {
     });
   }
 
+  componentWillReceiveProps(nextProps) {
+    this.props.fetchApiQuery.refetch();
+  }
+
   componentWillUnmount() {
     this.unsubscribe();
   }
@@ -46,7 +49,6 @@ class EventContainer extends React.Component<FinalProps, {}> {
     const {
       fetchApiQuery,
       removeEventMutation,
-      accountId,
       startTime,
       endTime,
       queryParams
@@ -63,12 +65,12 @@ class EventContainer extends React.Component<FinalProps, {}> {
     }
 
     // remove action
-    const remove = (event: IEvent) => {
+    const remove = (_id: string, accountId: string) => {
       confirm(getWarningMessage('Event'), { hasDeleteConfirm: true }).then(
         () => {
           removeEventMutation({
             variables: {
-              _id: event.providerEventId,
+              _id,
               accountId
             }
           })
@@ -100,12 +102,12 @@ export default withProps<Props>(
   compose(
     graphql<Props, any>(gql(queries.fetchApi), {
       name: 'fetchApiQuery',
-      options: ({ startTime, endTime, queryParams }) => {
+      options: ({ startTime, endTime, calendarIds }) => {
         return {
           variables: {
             path: '/nylas/get-events',
             params: {
-              ...queryParams,
+              calendarIds,
               startTime,
               endTime
             }
