@@ -1,18 +1,13 @@
 import gql from 'graphql-tag';
-import * as compose from 'lodash.flowright';
 import ButtonMutate from 'modules/common/components/ButtonMutate';
-import Spinner from 'modules/common/components/Spinner';
 import { IButtonMutateProps } from 'modules/common/types';
-import { withProps } from 'modules/common/utils';
 import { queries } from 'modules/settings/integrations/graphql';
 import React from 'react';
-import { graphql } from 'react-apollo';
 import EventForm from '../components/EventForm';
 import { mutations } from '../graphql';
-import { IEvent } from '../types';
+import { IAccount, IEvent } from '../types';
 
 type Props = {
-  accountId: string;
   queryParams: any;
   startTime?: Date;
   endTime?: Date;
@@ -20,32 +15,12 @@ type Props = {
   onHideModal: (date?: Date) => void;
   selectedDate?: Date;
   event?: IEvent;
+  account?: IAccount;
 };
 
-type FinalProps = {
-  fetchApiQuery: any;
-} & Props;
-
-class FormContainer extends React.Component<FinalProps, {}> {
+class FormContainer extends React.Component<Props> {
   render() {
-    const {
-      fetchApiQuery,
-      queryParams,
-      accountId,
-      startTime,
-      endTime,
-      event
-    } = this.props;
-
-    if (fetchApiQuery.loading) {
-      return <Spinner objective={true} />;
-    }
-
-    if (fetchApiQuery.error) {
-      return (
-        <span style={{ color: 'red' }}>Integrations api is not running</span>
-      );
-    }
+    const { queryParams, startTime, endTime, event } = this.props;
 
     const refetchQuery =
       startTime && endTime
@@ -70,8 +45,7 @@ class FormContainer extends React.Component<FinalProps, {}> {
       };
 
       const variables = {
-        ...values,
-        accountId
+        ...values
       };
 
       if (event) {
@@ -94,7 +68,6 @@ class FormContainer extends React.Component<FinalProps, {}> {
 
     const updatedProps = {
       ...this.props,
-      calendars: fetchApiQuery.integrationsFetchApi || [],
       renderButton
     };
 
@@ -102,20 +75,4 @@ class FormContainer extends React.Component<FinalProps, {}> {
   }
 }
 
-export default withProps<Props>(
-  compose(
-    graphql<Props, any>(gql(queries.fetchApi), {
-      name: 'fetchApiQuery',
-      options: ({ accountId }) => {
-        return {
-          variables: {
-            path: '/nylas/get-calendars',
-            params: {
-              accountId
-            }
-          }
-        };
-      }
-    })
-  )(FormContainer)
-);
+export default FormContainer;
