@@ -6,15 +6,15 @@ import { IRouterProps } from 'modules/common/types';
 import { router as routerUtils, withProps } from 'modules/common/utils';
 import { queries } from 'modules/settings/calendars/graphql';
 import {
-  GroupDetailQueryResponse,
-  GroupGetLastQueryResponse,
-  GroupsQueryResponse
+  BoardDetailQueryResponse,
+  BoardGetLastQueryResponse,
+  BoardsQueryResponse
 } from 'modules/settings/calendars/types';
 import queryString from 'query-string';
 import React from 'react';
 import { graphql } from 'react-apollo';
 import { withRouter } from 'react-router-dom';
-import Calendar from '../components/Calendar';
+import Calendar from '../components/Main';
 import { STORAGE_CALENDAR_GROUP_KEY, STORAGE_CALENDAR_KEY } from '../constants';
 
 type Props = {
@@ -23,104 +23,104 @@ type Props = {
 } & IRouterProps;
 
 type FinalProps = {
-  groupsQuery: GroupsQueryResponse;
-  groupGetLastQuery?: GroupGetLastQueryResponse;
-  groupDetailQuery?: GroupDetailQueryResponse;
+  boardsQuery: BoardsQueryResponse;
+  boardGetLastQuery?: BoardGetLastQueryResponse;
+  boardDetailQuery?: BoardDetailQueryResponse;
 } & Props;
 
 const generateQueryParams = ({ location }) => {
   return queryString.parse(location.search);
 };
 
-export const getGroupId = ({ location }) => {
+export const getBoardId = ({ location }) => {
   const queryParams = generateQueryParams({ location });
   return queryParams.id;
 };
 
 /*
- * Main group component
+ * Main board component
  */
 class Main extends React.Component<FinalProps> {
   render() {
     const {
       history,
       location,
-      groupsQuery,
-      groupGetLastQuery,
-      groupDetailQuery
+      boardsQuery,
+      boardGetLastQuery,
+      boardDetailQuery
     } = this.props;
 
-    if (groupsQuery.loading) {
+    if (boardsQuery.loading) {
       return <PageHeader />;
     }
 
     const queryParams = generateQueryParams({ location });
-    const groupId = getGroupId({ location });
-    const { calendarId } = queryParams;
+    const boardId = getBoardId({ location });
+    const { groupId } = queryParams;
 
-    if (groupId && calendarId) {
-      localStorage.setItem(STORAGE_CALENDAR_GROUP_KEY, groupId);
-      localStorage.setItem(STORAGE_CALENDAR_KEY, calendarId);
+    if (boardId && groupId) {
+      localStorage.setItem(STORAGE_CALENDAR_GROUP_KEY, boardId);
+      localStorage.setItem(STORAGE_CALENDAR_KEY, groupId);
     }
 
     // wait for load
-    if (groupDetailQuery && groupDetailQuery.loading) {
+    if (boardDetailQuery && boardDetailQuery.loading) {
       return <Spinner />;
     }
 
-    if (groupGetLastQuery && groupGetLastQuery.loading) {
+    if (boardGetLastQuery && boardGetLastQuery.loading) {
       return <Spinner />;
     }
 
-    const lastGroup =
-      groupGetLastQuery && groupGetLastQuery.calendarGroupGetLast;
-    const currentGroup =
-      groupDetailQuery && groupDetailQuery.calendarGroupDetail;
+    const lastBoard =
+      boardGetLastQuery && boardGetLastQuery.calendarBoardGetLast;
+    const currentBoard =
+      boardDetailQuery && boardDetailQuery.calendarBoardDetail;
 
-    // if there is no groupId in queryparams and there is one in localstorage
+    // if there is no boardId in queryparams and there is one in localstorage
     // then put those in queryparams
-    const defaultGroupId = localStorage.getItem(STORAGE_CALENDAR_GROUP_KEY);
-    const defaultCalendarId = localStorage.getItem(STORAGE_CALENDAR_KEY);
+    const defaultBoardId = localStorage.getItem(STORAGE_CALENDAR_GROUP_KEY);
+    const defaultGroupId = localStorage.getItem(STORAGE_CALENDAR_KEY);
 
-    if (!groupId && defaultGroupId) {
+    if (!boardId && defaultBoardId) {
       routerUtils.setParams(history, {
-        id: defaultGroupId,
-        calendarId: defaultCalendarId
+        id: defaultBoardId,
+        groupId: defaultGroupId
       });
 
       return null;
     }
 
-    // if there is no groupId in queryparams and there is lastGroup
-    // then put lastGroup._id and this group's first calendarId to queryparams
+    // if there is no boardId in queryparams and there is lastBoard
+    // then put lastBoard._id and this board's first groupId to queryparams
     if (
-      !groupId &&
-      lastGroup &&
-      lastGroup.calendars &&
-      lastGroup.calendars.length > 0
+      !boardId &&
+      lastBoard &&
+      lastBoard.groups &&
+      lastBoard.groups.length > 0
     ) {
-      const [firstCalendar] = lastGroup.calendars;
+      const [firstGroup] = lastBoard.groups;
 
       routerUtils.setParams(history, {
-        id: lastGroup._id,
-        calendarId: firstCalendar._id
+        id: lastBoard._id,
+        groupId: firstGroup._id
       });
 
       return null;
     }
 
-    const calendars = currentGroup ? currentGroup.calendars || [] : [];
+    const groups = currentBoard ? currentBoard.groups || [] : [];
 
-    const currentCalendar = calendarId
-      ? calendars.find(calendar => calendar._id === calendarId) || calendars[0]
-      : calendars[0];
+    const currentGroup = groupId
+      ? groups.find(group => group._id === groupId) || groups[0]
+      : groups[0];
 
     const props = {
       queryParams,
       history,
+      currentBoard,
       currentGroup,
-      currentCalendar,
-      groups: groupsQuery.calendarGroups || []
+      boards: boardsQuery.calendarBoards || []
     };
 
     const extendedProps = {
@@ -133,26 +133,26 @@ class Main extends React.Component<FinalProps> {
 
 const MainActionBarContainer = withProps<Props>(
   compose(
-    graphql<Props, GroupsQueryResponse>(gql(queries.groups), {
-      name: 'groupsQuery',
+    graphql<Props, BoardsQueryResponse>(gql(queries.boards), {
+      name: 'boardsQuery',
       options: () => ({
         variables: {}
       })
     }),
-    graphql<Props, GroupGetLastQueryResponse>(gql(queries.groupGetLast), {
-      name: 'groupGetLastQuery',
-      skip: props => getGroupId(props),
+    graphql<Props, BoardGetLastQueryResponse>(gql(queries.boardGetLast), {
+      name: 'boardGetLastQuery',
+      skip: props => getBoardId(props),
       options: () => ({
         variables: {}
       })
     }),
-    graphql<Props, GroupDetailQueryResponse, { _id: string }>(
-      gql(queries.groupDetail),
+    graphql<Props, BoardDetailQueryResponse, { _id: string }>(
+      gql(queries.boardDetail),
       {
-        name: 'groupDetailQuery',
-        skip: props => !getGroupId(props),
+        name: 'boardDetailQuery',
+        skip: props => !getBoardId(props),
         options: props => ({
-          variables: { _id: getGroupId(props) }
+          variables: { _id: getBoardId(props) }
         })
       }
     )
