@@ -1,4 +1,5 @@
 import mutations from './mutations';
+import queries from './queries';
 import subscriptions from './subscriptions';
 
-export { mutations, subscriptions };
+export { mutations, subscriptions, queries };
