@@ -3,6 +3,25 @@ import Spinner from 'modules/common/components/Spinner';
 import React from 'react';
 import { IAnimatedLoader } from '../types';
 
+function retry(fn, retriesLeft = 30, interval = 2000) {
+  return new Promise((resolve, reject) => {
+    fn()
+      .then(resolve)
+      .catch(error => {
+        setTimeout(() => {
+          if (retriesLeft === 1) {
+            // reject('maximum retries exceeded');
+            reject(error);
+            return;
+          }
+
+          // Passing on "reject" is the important part
+          retry(fn, retriesLeft - 1, interval).then(resolve, reject);
+        }, interval);
+      });
+  });
+}
+
 export default function asyncComponent(
   importComponent: any,
   loaderStyle?: IAnimatedLoader
@@ -17,7 +36,7 @@ export default function asyncComponent(
     }
 
     async componentDidMount() {
-      const { default: component } = await importComponent();
+      const { default: component }: any = await retry(() => importComponent());
 
       this.setState({ component });
     }
