@@ -13,7 +13,14 @@ const createVideoChatRoom = `
   }
 `;
 
+const saveVideoRecordingInfo = `
+  mutation conversationsSaveVideoRecordingInfo($conversationId: String!, $recordingId: String!) {
+    conversationsSaveVideoRecordingInfo(conversationId: $conversationId, recordingId: $recordingId)
+  }
+`;
+
 export default {
   deleteVideoChatRoom,
-  createVideoChatRoom
+  createVideoChatRoom,
+  saveVideoRecordingInfo
 };
