@@ -30,6 +30,7 @@ interface IState {
   isBrowserInfoSaved: boolean;
   headHeight: number;
   botTyping: boolean;
+  browserInfo: IBrowserInfo;
 }
 
 interface IStore extends IState {
@@ -56,7 +57,7 @@ interface IStore extends IState {
   readConversation: (conversationId: string) => void;
   readMessages: (conversationId: string) => void;
   replyAutoAnswer: (message: string, payload: string, type: string) => void;
-  getBotInitialMessage: () => void;
+  getBotInitialMessage: (callback: (bodData: any) => void) => void;
   changeOperatorStatus: (
     _id: string,
     operatorStatus: string,
@@ -74,6 +75,7 @@ interface IStore extends IState {
   isLoggedIn: () => boolean;
   setBotTyping: (typing: boolean) => void;
   botTyping: boolean;
+  browserInfo: IBrowserInfo;
 }
 
 export const MESSAGE_TYPES = {
@@ -122,7 +124,8 @@ export class AppProvider extends React.Component<{}, IState> {
       isAttachingFile: false,
       isBrowserInfoSaved: false,
       headHeight: 200,
-      botTyping: false
+      botTyping: false,
+      browserInfo: {}
     };
   }
 
@@ -180,7 +183,8 @@ export class AppProvider extends React.Component<{}, IState> {
           .then(({ data: { widgetsSaveBrowserInfo } }: any) => {
             this.setState({
               lastUnreadMessage: widgetsSaveBrowserInfo,
-              isBrowserInfoSaved: true
+              isBrowserInfoSaved: true,
+              browserInfo
             });
           });
       }
@@ -399,10 +403,7 @@ export class AppProvider extends React.Component<{}, IState> {
   };
 
   readMessages = (conversationId: string) => {
-    if (this.state.unreadCount === 0) {
-      return;
-    }
-
+    
     client
       .mutate({
         mutation: gql(graphqlTypes.readConversationMessages),
@@ -469,27 +470,24 @@ export class AppProvider extends React.Component<{}, IState> {
       });
   };
 
-  getBotInitialMessage = () => {
+  getBotInitialMessage = (callback: (botData: any) => void) => {
     return client.mutate({
       mutation: gql`
         mutation widgetGetBotInitialMessage(
-          $customerId: String,
           $integrationId: String
         ) {
             widgetGetBotInitialMessage(
-            customerId: $customerId
             integrationId: $integrationId
           )
         }
       `,
       variables: {
         integrationId: connection.data.integrationId,
-        customerId: connection.data.customerId,
       }
     })
       .then(({ data }) => {
         if (data.widgetGetBotInitialMessage) {
-          this.setState({ activeConversation: data.widgetGetBotInitialMessage })
+          callback(data.widgetGetBotInitialMessage);
         }
       })
   };
@@ -504,7 +502,7 @@ export class AppProvider extends React.Component<{}, IState> {
             $message: String!
             $payload: String!
             $type: String!
-            $conversationId: String!
+            $conversationId: String
             $customerId: String!
             $integrationId: String!
           ) {
@@ -527,8 +525,13 @@ export class AppProvider extends React.Component<{}, IState> {
           payload
         }
       })
-      .then(() => {
-        this.setState({ sendingMessage: false });
+      .then(({ data }) => {
+        const { conversationId } = data.widgetBotRequest;
+
+        this.setState({
+          sendingMessage: false,
+          activeConversation: conversationId
+        });
       })
       .catch(() => {
         this.setState({ sendingMessage: false });
@@ -721,7 +724,8 @@ export class AppProvider extends React.Component<{}, IState> {
           sendFile: this.sendFile,
           setHeadHeight: this.setHeadHeight,
           setUnreadCount: this.setUnreadCount,
-          isLoggedIn: this.isLoggedIn
+          isLoggedIn: this.isLoggedIn,
+          browserInfo: this.state.browserInfo,
         }}
       >
         {this.props.children}
