@@ -1,16 +1,22 @@
+import gql from "graphql-tag";
 import * as React from "react";
-
+import { compose, graphql } from "react-apollo";
+import { IBrowserInfo } from "../../types";
 import { Notifier as DumbNotifier } from "../components";
-import { IMessage } from "../types";
+import { connection } from '../connection';
+import graphqlTypes from '../graphql';
+import { EngageMessageQueryResponse ,IMessage} from "../types";
 import { AppConsumer } from "./AppContext";
-
 type Props = {
   message?: IMessage;
+  browserInfo?: IBrowserInfo;
+  engageMessageQuery: EngageMessageQueryResponse;
 };
-
-export default class Notifier extends React.Component<Props> {
+class Notifier extends React.Component<Props> {
   render() {
-    const { message } = this.props;
+    const { engageMessageQuery} = this.props;
+
+    const message = engageMessageQuery.widgetsGetEngageMessage;
 
     if (!message || !message._id) {
       return null;
@@ -45,3 +51,24 @@ export default class Notifier extends React.Component<Props> {
     );
   }
 }
+
+
+const withPollInterval = compose(graphql<Props>(
+  gql(graphqlTypes.getEngageMessage), {
+  name: 'engageMessageQuery',
+  options: ownProps => ({
+    variables: {
+      customerId: connection.data.customerId,
+      browserInfo: ownProps.browserInfo,
+    },
+    notifyOnNetworkStatusChange: true,
+    fetchPolicy: 'network-only',
+    skip: !connection.data.customerId,
+    // every minute
+    pollInterval: 60000
+  })
+})
+)(Notifier)
+
+
+export default withPollInterval
